requirements.txt
flask==3.0.0
anthropic==0.40.0
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-flask==1.3.0
